+++ include/coord.h
@@ -42,6 +42,7 @@ extern int seekcoord(Coordfile,int);
 extern Bool readcoord(Coordfile,Coord *,const Coord *);
 extern void closecoord(Coordfile);
 extern Bool bstruct_writecoord(Bstruct,const char *,const Coord *);
+extern Bool bstruct_readcoord(Bstruct,const char *,Coord *);
 extern Bool writecoord(FILE *,const Coord *);
 extern Bool writefloatcoord(FILE *,const Coord *);
 extern Real cellarea(const Coord *,const Coord *);