+++ src/test/support/header_of_lpjml_files_to_link/bstruct_getnoread.h
@@ -1,0 +1,1 @@
+int bstruct_getnoread(const Bstruct);