+++ src/test/support/header_of_lpjml_files_to_link/bstruct_readid.h
@@ -1,0 +1,1 @@
+Bool bstruct_readid(Bstruct,Byte,Id *);